+++ spec/unit/resources/command_line_tools_spec.rb
@@ -21,10 +21,11 @@ describe 'command_line_tools' do
       end
     end
 
-    it { is_expected.to create_file('create sentinel file') }
-    it { is_expected.to run_execute('install Command Line Tools (macOS High Sierra version 10.13) for Xcode-10.0') }
-    it { is_expected.to delete_file('delete sentinel file') }
+    it { is_expected.to create_file('create demand file') }
+    it { expect(chef_run.file('create demand file')).to notify('execute[install command line tools]').to(:run).immediately }
+    it { is_expected.to delete_file('delete demand file') }
   end
+
 
   context 'with libxcrun present' do
     before do
@@ -37,9 +38,8 @@ describe 'command_line_tools' do
       end
     end
 
-    it { is_expected.to create_file('create sentinel file') }
-    it { is_expected.to_not run_execute('install Command Line Tools (macOS High Sierra version 10.13) for Xcode-10.0') }
-    it { is_expected.to delete_file('delete sentinel file') }
+    it { is_expected.to_not create_file('create demand file') }
+    it { is_expected.to_not run_execute('install command line tools') }
   end
 end
 
@@ -65,11 +65,11 @@ describe 'command_line_tools' do
       end
     end
 
-    it { is_expected.to create_file('create sentinel file') }
+    it { is_expected.to create_file('create demand file') }
     it {
       is_expected.to run_execute('upgrade Command Line Tools for Xcode-11.0')
         .with(command: ['softwareupdate', '--install', 'Command Line Tools for Xcode-22.0'])
     }
-    it { is_expected.to delete_file('delete sentinel file') }
+    it { is_expected.to delete_file('delete demand file') }
   end
 end